legistra/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.0.354
transformers==4.35.2
torch==2.1.1
faiss-cpu==1.7.4
reportlab==4.0.7
celery==5.3.4
redis==5.0.1
pytest==7.4.0
datasets==2.14.5
accelerate==0.22.0
peft==0.4.0
mlflow==2.13.0
evaluate
wandb
dvc
evidently
scikit-learn
nltk
sentence-transformers
